Move all CGH tests that do not require runtime to handler_tests.cc
